+++ packages/tsurlfilter/CHANGELOG.md
@@ -8,6 +8,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 <!-- TODO: manually add compare links for version changes -->
 <!-- e.g. [1.0.77]: https://github.com/AdguardTeam/tsurlfilter/compare/tsurlfilter-v1.0.76...tsurlfilter-v1.0.77 -->
 
+## [2.2.16] - 2024-03-26
+
+### Fixed
+
+- Correct work `$all` and `$popup` when they both selected for request [#2620],
+  [#2728].
+
+[2.2.16]: https://github.com/AdguardTeam/tsurlfilter/releases/tag/tsurlfilter-v2.2.16
+[#2620]: https://github.com/AdguardTeam/AdguardBrowserExtension/issues/2620
+[#2728]: https://github.com/AdguardTeam/AdguardBrowserExtension/issues/2728
+
+
 ## [2.2.15] - 2024-03-01
 
 ### Changed